+++ History.md
@@ -2,6 +2,8 @@ Change Log
 ----
 ### 0.6.5
 - Add archiving of tasks to existing archive or deleted list or create archive list
+- Add filter by selected tasks filenames
+- Allow regex in filter
 
 ### 0.6.4
 - Add batch move tasks